+++ lib/src/plugin_method.dart
@@ -1,0 +1,10 @@
+enum PluginMethod {
+  logIn,
+  expressLogIn,
+  logOut,
+  getAccessToken,
+  getUserProfile,
+  getUserEmail,
+  getProfileImageUrl,
+  getSdkVersion,
+}